Factor out `push` from `device`
