+++ components/fxa-client/src/lib.rs
@@ -42,18 +42,17 @@ mod device;
 mod error;
 mod migration;
 mod profile;
+mod push;
 mod storage;
 mod telemetry;
 mod token;
 
 pub use auth::{AuthorizationInfo, MetricsParams};
-pub use device::{
-    AccountEvent, AttachedClient, Device, DeviceCapability, DevicePushSubscription,
-    IncomingDeviceCommand, SendTabPayload, TabHistoryEntry,
-};
+pub use device::{AttachedClient, Device, DeviceCapability};
 pub use error::{FxaError, Error};
 pub use migration::{FxAMigrationResult, MigrationState};
 pub use profile::Profile;
+pub use push::{AccountEvent, DevicePushSubscription, IncomingDeviceCommand, SendTabPayload, TabHistoryEntry};
 pub use token::{AccessTokenInfo, AuthorizationParameters, ScopedKey};
 
 // All the implementation details live in this "internal" module.
